refactor: split 4 moderate oversized tool files into focused modules
Split partman/operations.ts (1013 lines):
- helpers.ts (67 lines): shared utilities (getPartmanSchema, ensurePartmanSchemaAlias, callPartmanProcedure)
- operations.ts (490 lines): check_default, partition_data tools
- maintenance.ts (638 lines): set_retention, undo_partition, analyze_health tools

Split core/convenience.ts (856 lines):
- convenience.ts (687 lines): CRUD schemas + upsert, batch_insert tools
- utility.ts (204 lines): count, exists, truncate tools + getConvenienceTools

Split citext.ts (846 lines) -> citext/ directory:
- setup.ts (241 lines): create extension, convert column tools
- analysis.ts (552 lines): list, analyze, compare, advisor tools
- index.ts (47 lines): barrel file

Split partitioning.ts (845 lines) -> partitioning/ directory:
- management.ts (487 lines): helpers + list, create table, create partition tools
- info.ts (362 lines): attach, detach, partition info tools
- index.ts (49 lines): barrel file

Updated barrel files, PostgresAdapter.ts imports, and test file imports.
