+++ src/adapters/postgresql/PostgresAdapter.ts
@@ -43,12 +43,12 @@ import { getBackupTools } from "./tools/backup/index.js";
 import { getSchemaTools } from "./tools/schema/index.js";
 import { getVectorTools } from "./tools/vector/index.js";
 import { getPostgisTools } from "./tools/postgis/index.js";
-import { getPartitioningTools } from "./tools/partitioning.js";
+import { getPartitioningTools } from "./tools/partitioning/index.js";
 import { getStatsTools } from "./tools/stats/index.js";
 import { getCronTools } from "./tools/cron.js";
 import { getPartmanTools } from "./tools/partman/index.js";
 import { getKcacheTools } from "./tools/kcache.js";
-import { getCitextTools } from "./tools/citext.js";
+import { getCitextTools } from "./tools/citext/index.js";
 import { getLtreeTools } from "./tools/ltree.js";
 import { getPgcryptoTools } from "./tools/pgcrypto.js";
 import { getIntrospectionTools } from "./tools/introspection/index.js";
