+++ src/adapters/postgresql/tools/__tests__/partitioning.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { describe, it, expect, vi, beforeEach } from "vitest";
-import { getPartitioningTools } from "../partitioning.js";
+import { getPartitioningTools } from "../partitioning/index.js";
 import type { PostgresAdapter } from "../../PostgresAdapter.js";
 import {
   createMockPostgresAdapter,
