+++ src/adapters/postgresql/tools/partitioning/management.ts
@@ -1,48 +1,39 @@
 /**
- * PostgreSQL Partitioning Tools
+ * PostgreSQL Partitioning Tools - Management
  *
- * Table partitioning management.
- * 6 tools total.
+ * Partition management: list, create table, create partition, attach.
  */
 
-import type { PostgresAdapter } from "../PostgresAdapter.js";
-import type { ToolDefinition, RequestContext } from "../../../types/index.js";
-import { readOnly, write, destructive } from "../../../utils/annotations.js";
-import { formatPostgresError } from "./core/error-helpers.js";
-import { getToolIcons } from "../../../utils/icons.js";
+import type { PostgresAdapter } from "../../PostgresAdapter.js";
+import type {
+  ToolDefinition,
+  RequestContext,
+} from "../../../../types/index.js";
+
+import { readOnly, write } from "../../../../utils/annotations.js";
+import { getToolIcons } from "../../../../utils/icons.js";
+import { formatPostgresError } from "../core/error-helpers.js";
 import {
   sanitizeIdentifier,
   sanitizeTableName,
-} from "../../../utils/identifiers.js";
+} from "../../../../utils/identifiers.js";
 import {
-  // Base schemas for MCP visibility
-  CreatePartitionedTableSchemaBase,
-  CreatePartitionSchemaBase,
-  AttachPartitionSchemaBase,
-  DetachPartitionSchemaBase,
-  ListPartitionsSchemaBase,
-  PartitionInfoSchemaBase,
-  // Preprocessed schemas for handler parsing
   CreatePartitionedTableSchema,
+  CreatePartitionedTableSchemaBase,
   CreatePartitionSchema,
-  AttachPartitionSchema,
-  DetachPartitionSchema,
+  CreatePartitionSchemaBase,
   ListPartitionsSchema,
-  PartitionInfoSchema,
-  // Output schemas
+  ListPartitionsSchemaBase,
   ListPartitionsOutputSchema,
   CreatePartitionedTableOutputSchema,
   CreatePartitionOutputSchema,
-  AttachPartitionOutputSchema,
-  DetachPartitionOutputSchema,
-  PartitionInfoOutputSchema,
-} from "../schemas/index.js";
-
+} from "../../schemas/index.js";
 /**
  * Parse schema.table format identifier
  * Returns { table, schema } with schema extracted from prefix if present
  */
-function parseSchemaTable(
+
+export function parseSchemaTable(
   identifier: string,
   defaultSchema?: string,
 ): { table: string; schema: string } {
@@ -59,7 +50,7 @@ function parseSchemaTable(
 /**
  * Format bytes to human-readable string with consistent formatting
  */
-function formatBytes(bytes: number): string {
+export function formatBytes(bytes: number): string {
   if (bytes < 1024) return `${String(bytes)} B`;
   if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
   if (bytes < 1024 * 1024 * 1024)
@@ -71,15 +62,15 @@ function formatBytes(bytes: number): string {
  * Check table existence and partition status
  * Returns: 'partitioned' | 'not_partitioned' | 'not_found'
  */
-async function checkTablePartitionStatus(
+export async function checkTablePartitionStatus(
   adapter: PostgresAdapter,
   table: string,
   schema: string,
 ): Promise<"partitioned" | "not_partitioned" | "not_found"> {
   // 'r' = regular table, 'p' = partitioned table
-  const checkSql = `SELECT c.relkind FROM pg_class c 
+  const checkSql = `SELECT c.relkind FROM pg_class c
         JOIN pg_namespace n ON c.relnamespace = n.oid
-        WHERE c.relname = $1 AND n.nspname = $2 
+        WHERE c.relname = $1 AND n.nspname = $2
         AND c.relkind IN ('r', 'p')`;
   const result = await adapter.executeQuery(checkSql, [table, schema]);
 
@@ -91,23 +82,9 @@ async function checkTablePartitionStatus(
   return rows[0]?.["relkind"] === "p" ? "partitioned" : "not_partitioned";
 }
 
-/**
- * Get all partitioning tools
- */
-export function getPartitioningTools(
+export function createListPartitionsTool(
   adapter: PostgresAdapter,
-): ToolDefinition[] {
-  return [
-    createListPartitionsTool(adapter),
-    createPartitionedTableTool(adapter),
-    createPartitionTool(adapter),
-    createAttachPartitionTool(adapter),
-    createDetachPartitionTool(adapter),
-    createPartitionInfoTool(adapter),
-  ];
-}
-
-function createListPartitionsTool(adapter: PostgresAdapter): ToolDefinition {
+): ToolDefinition {
   return {
     name: "pg_list_partitions",
     description:
@@ -162,7 +139,7 @@ function createListPartitionsTool(adapter: PostgresAdapter): ToolDefinition {
       const limit = parsed.limit ?? 50;
 
       // Build query with optional limit
-      let sql = `SELECT 
+      let sql = `SELECT
                         c.relname as partition_name,
                         pg_get_expr(c.relpartbound, c.oid) as bounds,
                         pg_table_size(c.oid) as size_bytes,
@@ -220,7 +197,9 @@ function createListPartitionsTool(adapter: PostgresAdapter): ToolDefinition {
   };
 }
 
-function createPartitionedTableTool(adapter: PostgresAdapter): ToolDefinition {
+export function createPartitionedTableTool(
+  adapter: PostgresAdapter,
+): ToolDefinition {
   return {
     name: "pg_create_partitioned_table",
     description:
@@ -389,7 +368,7 @@ function createPartitionedTableTool(adapter: PostgresAdapter): ToolDefinition {
   };
 }
 
-function createPartitionTool(adapter: PostgresAdapter): ToolDefinition {
+export function createPartitionTool(adapter: PostgresAdapter): ToolDefinition {
   return {
     name: "pg_create_partition",
     description:
@@ -507,338 +486,6 @@ function createPartitionTool(adapter: PostgresAdapter): ToolDefinition {
       }
 
       return result;
-    },
-  };
-}
-
-function createAttachPartitionTool(adapter: PostgresAdapter): ToolDefinition {
-  return {
-    name: "pg_attach_partition",
-    description: "Attach an existing table as a partition.",
-    group: "partitioning",
-    inputSchema: AttachPartitionSchemaBase, // Base schema for MCP visibility
-    outputSchema: AttachPartitionOutputSchema,
-    annotations: write("Attach Partition"),
-    icons: getToolIcons("partitioning", write("Attach Partition")),
-    handler: async (params: unknown, _context: RequestContext) => {
-      const { parent, partition, forValues, schema } =
-        AttachPartitionSchema.parse(params) as {
-          parent: string;
-          partition: string;
-          forValues: string;
-          schema?: string;
-        };
-
-      // Check parent table existence and partition status before SQL execution
-      const parsedParentCheck = parseSchemaTable(parent, schema);
-      const parentStatus = await checkTablePartitionStatus(
-        adapter,
-        parsedParentCheck.table,
-        parsedParentCheck.schema,
-      );
-      if (parentStatus === "not_found") {
-        return {
-          success: false,
-          error: `Parent table '${parsedParentCheck.schema}.${parsedParentCheck.table}' does not exist.`,
-        };
-      }
-      if (parentStatus === "not_partitioned") {
-        return {
-          success: false,
-          error: `Parent table '${parsedParentCheck.schema}.${parsedParentCheck.table}' exists but is not partitioned.`,
-        };
-      }
-
-      // Check partition table exists (it must exist as a standalone table to attach)
-      const parsedPartCheck = parseSchemaTable(partition, schema);
-      const partCheckSql = `SELECT 1 FROM pg_class c JOIN pg_namespace n ON c.relnamespace = n.oid WHERE c.relname = $1 AND n.nspname = $2 AND c.relkind IN ('r', 'p')`;
-      const partCheckResult = await adapter.executeQuery(partCheckSql, [
-        parsedPartCheck.table,
-        parsedPartCheck.schema,
-      ]);
-      if ((partCheckResult.rows ?? []).length === 0) {
-        return {
-          success: false,
-          error: `Partition table '${parsedPartCheck.schema}.${parsedPartCheck.table}' does not exist.`,
-        };
-      }
-
-      // Parse schema.table format from parent and partition (takes priority over explicit schema)
-      const parsedParent = parseSchemaTable(parent, schema);
-      const parsedPartition = parseSchemaTable(partition, schema);
-
-      // Use parent's schema if partition doesn't have schema prefix and no explicit schema
-      const resolvedPartitionSchema = partition.includes(".")
-        ? parsedPartition.schema
-        : (schema ?? parsedParent.schema);
-
-      const parentName = sanitizeTableName(
-        parsedParent.table,
-        parsedParent.schema,
-      );
-      const partitionName = sanitizeTableName(
-        parsedPartition.table,
-        resolvedPartitionSchema,
-      );
-
-      // Handle DEFAULT partition
-      // Accept both "__DEFAULT__" (from preprocessor when isDefault: true) and explicit "DEFAULT"
-      const isDefaultPartition =
-        forValues === "__DEFAULT__" ||
-        forValues.toUpperCase() === "DEFAULT" ||
-        forValues.toUpperCase().trim() === "DEFAULT";
-
-      let sql: string;
-      let boundsDescription: string;
-      if (isDefaultPartition) {
-        sql = `ALTER TABLE ${parentName} ATTACH PARTITION ${partitionName} DEFAULT`;
-        boundsDescription = "DEFAULT";
-      } else {
-        sql = `ALTER TABLE ${parentName} ATTACH PARTITION ${partitionName} FOR VALUES ${forValues}`;
-        boundsDescription = forValues;
-      }
-
-      try {
-        await adapter.executeQuery(sql);
-      } catch (error: unknown) {
-        return {
-          success: false,
-          error: formatPostgresError(error, {
-            tool: "pg_attach_partition",
-            table: parsedPartition.table,
-          }),
-        };
-      }
-
-      return {
-        success: true,
-        parent: parsedParent.table,
-        partition: parsedPartition.table,
-        bounds: boundsDescription,
-      };
-    },
-  };
-}
-
-function createDetachPartitionTool(adapter: PostgresAdapter): ToolDefinition {
-  return {
-    name: "pg_detach_partition",
-    description:
-      "Detach a partition. Use concurrently: true for non-blocking. Use finalize: true only after an interrupted CONCURRENTLY detach.",
-    group: "partitioning",
-    inputSchema: DetachPartitionSchemaBase, // Base schema for MCP visibility
-    outputSchema: DetachPartitionOutputSchema,
-    annotations: destructive("Detach Partition"),
-    icons: getToolIcons("partitioning", destructive("Detach Partition")),
-    handler: async (params: unknown, _context: RequestContext) => {
-      const { parent, partition, concurrently, finalize, schema } =
-        DetachPartitionSchema.parse(params) as {
-          parent: string;
-          partition: string;
-          concurrently?: boolean;
-          finalize?: boolean;
-          schema?: string;
-        };
-
-      // Check parent table existence and partition status before SQL execution
-      const parsedParentCheck = parseSchemaTable(parent, schema);
-      const parentStatus = await checkTablePartitionStatus(
-        adapter,
-        parsedParentCheck.table,
-        parsedParentCheck.schema,
-      );
-      if (parentStatus === "not_found") {
-        return {
-          success: false,
-          error: `Parent table '${parsedParentCheck.schema}.${parsedParentCheck.table}' does not exist.`,
-        };
-      }
-      if (parentStatus === "not_partitioned") {
-        return {
-          success: false,
-          error: `Parent table '${parsedParentCheck.schema}.${parsedParentCheck.table}' exists but is not partitioned.`,
-        };
-      }
-
-      // Check partition table exists
-      const parsedPartCheck = parseSchemaTable(partition, schema);
-      const partCheckSql = `SELECT 1 FROM pg_class c JOIN pg_namespace n ON c.relnamespace = n.oid WHERE c.relname = $1 AND n.nspname = $2`;
-      const partCheckResult = await adapter.executeQuery(partCheckSql, [
-        parsedPartCheck.table,
-        parsedPartCheck.schema,
-      ]);
-      if ((partCheckResult.rows ?? []).length === 0) {
-        return {
-          success: false,
-          error: `Partition '${parsedPartCheck.schema}.${parsedPartCheck.table}' does not exist.`,
-        };
-      }
-
-      // Parse schema.table format from parent and partition (takes priority over explicit schema)
-      const parsedParent = parseSchemaTable(parent, schema);
-      const parsedPartition = parseSchemaTable(partition, schema);
-
-      // Use parent's schema if partition doesn't have schema prefix and no explicit schema
-      const resolvedPartitionSchema = partition.includes(".")
-        ? parsedPartition.schema
-        : (schema ?? parsedParent.schema);
-
-      const parentName = sanitizeTableName(
-        parsedParent.table,
-        parsedParent.schema,
-      );
-      const partitionName = sanitizeTableName(
-        parsedPartition.table,
-        resolvedPartitionSchema,
-      );
-
-      // Build the appropriate clause
-      let clause = "";
-      if (finalize === true) {
-        // FINALIZE is used to complete an interrupted CONCURRENTLY detach
-        clause = " FINALIZE";
-      } else if (concurrently === true) {
-        clause = " CONCURRENTLY";
-      }
-
-      const sql = `ALTER TABLE ${parentName} DETACH PARTITION ${partitionName}${clause}`;
-
-      try {
-        await adapter.executeQuery(sql);
-      } catch (error: unknown) {
-        return {
-          success: false,
-          error: formatPostgresError(error, {
-            tool: "pg_detach_partition",
-            table: parsedPartition.table,
-          }),
-        };
-      }
-
-      return {
-        success: true,
-        parent: parsedParent.table,
-        partition: parsedPartition.table,
-      };
-    },
-  };
-}
-
-function createPartitionInfoTool(adapter: PostgresAdapter): ToolDefinition {
-  return {
-    name: "pg_partition_info",
-    description:
-      "Get detailed information about a partitioned table. Returns warning if table is not partitioned.",
-    group: "partitioning",
-    inputSchema: PartitionInfoSchemaBase, // Base schema for MCP visibility with alias support
-    outputSchema: PartitionInfoOutputSchema,
-    annotations: readOnly("Partition Info"),
-    icons: getToolIcons("partitioning", readOnly("Partition Info")),
-    handler: async (params: unknown, _context: RequestContext) => {
-      // Use preprocessed schema for alias resolution
-      const parsed = PartitionInfoSchema.parse(params) as {
-        table: string;
-        schema?: string;
-      };
-
-      // Parse schema.table format if present
-      let tableName = parsed.table;
-      let schemaName = parsed.schema ?? "public";
-      if (tableName.includes(".")) {
-        const parts = tableName.split(".");
-        schemaName = parts[0] ?? "public";
-        tableName = parts[1] ?? tableName;
-      }
-
-      // Check table existence and partition status
-      const resolvedTable = tableName;
-      const tableStatus = await checkTablePartitionStatus(
-        adapter,
-        resolvedTable,
-        schemaName,
-      );
-      if (tableStatus === "not_found") {
-        return {
-          tableInfo: null,
-          partitions: [],
-          totalSizeBytes: 0,
-          warning: `Table '${schemaName}.${resolvedTable}' does not exist.`,
-        };
-      }
-      if (tableStatus === "not_partitioned") {
-        return {
-          tableInfo: null,
-          partitions: [],
-          totalSizeBytes: 0,
-          warning: `Table '${schemaName}.${resolvedTable}' exists but is not partitioned. Use pg_create_partitioned_table to create a partitioned table.`,
-        };
-      }
-
-      const partInfoSql = `SELECT 
-                        c.relname as table_name,
-                        CASE pt.partstrat 
-                            WHEN 'r' THEN 'RANGE'
-                            WHEN 'l' THEN 'LIST'
-                            WHEN 'h' THEN 'HASH'
-                        END as partition_strategy,
-                        pg_get_partkeydef(c.oid) as partition_key,
-                        (SELECT count(*) FROM pg_inherits WHERE inhparent = c.oid) as partition_count
-                        FROM pg_class c
-                        JOIN pg_partitioned_table pt ON c.oid = pt.partrelid
-                        JOIN pg_namespace n ON c.relnamespace = n.oid
-                        WHERE c.relname = $1 AND n.nspname = $2`;
-
-      const partInfo = await adapter.executeQuery(partInfoSql, [
-        resolvedTable,
-        schemaName,
-      ]);
-
-      const partitionsSql = `SELECT 
-                        c.relname as partition_name,
-                        pg_get_expr(c.relpartbound, c.oid) as bounds,
-                        pg_table_size(c.oid) as size_bytes,
-                        GREATEST(0, (SELECT reltuples::bigint FROM pg_class WHERE oid = c.oid)) as approx_rows
-                        FROM pg_class c
-                        JOIN pg_inherits i ON c.oid = i.inhrelid
-                        WHERE i.inhparent = ($1 || '.' || $2)::regclass
-                        ORDER BY c.relname`;
-
-      const partitionsResult = await adapter.executeQuery(partitionsSql, [
-        schemaName,
-        resolvedTable,
-      ]);
-
-      // Calculate total size before mapping
-      const totalSizeBytes = (partitionsResult.rows ?? []).reduce(
-        (sum, row) => sum + Number(row["size_bytes"] ?? 0),
-        0,
-      );
-
-      // Format sizes consistently and coerce numeric fields
-      const partitions = (partitionsResult.rows ?? []).map((row) => {
-        const sizeBytes = Number(row["size_bytes"] ?? 0);
-        return {
-          ...row,
-          size_bytes: sizeBytes,
-          size: formatBytes(sizeBytes),
-          approx_rows: Number(row["approx_rows"] ?? 0),
-        };
-      });
-
-      // Coerce tableInfo numeric fields
-      const tableInfoRaw = partInfo.rows?.[0];
-      const tableInfo = tableInfoRaw
-        ? {
-            ...tableInfoRaw,
-            partition_count: Number(tableInfoRaw["partition_count"] ?? 0),
-          }
-        : null;
-
-      return {
-        tableInfo,
-        partitions,
-        totalSizeBytes,
-      };
     },
   };
 }
